moved FilterMatch definition to its own file, so as to remove the circular dependency between filter.h and particle.h
