+++ include/openmc/geometry.h
@@ -7,8 +7,6 @@
 #include <vector>
 
 #include "openmc/particle.h"
-#include "openmc/tallies/filter.h"
-
 
 namespace openmc {
